habistats: Habitat Statistics

Summarize and visualize the statistics of spatial data. Ecology and evolutionary biology are the main targets of this package, and utility functions to handle IUCN Red List Spatial Data and Köppen-Geiger Climate Classification are provided. But the core part of the package is data-agnostic, and can be used with any spatial data, even if they are not related to habitat or biology.

See also

Author

Maintainer: Watal M. Iwasaki heavywatal@gmail.com (ORCID)

Authors:

Fuku Sakamoto fukusakamoto29@gmail.com (ORCID)