Use geometry column of IUCN Spatial Data

The geometry column of IUCN Spatial Data is evaluated in several ways. Currently, perimeter, area, and some other features are calculated.

Usage

evaluate_iucn_range(sci_name, iucn_source = NULL)

Arguments

sci_name

A character vector of species names.

iucn_source

Path to the directory containing IUCN shape files. iucn_source_path() is used if NULL is given (default).

Value

evaluate_iucn_range() returns a data.frame.

Examples

if (FALSE) { # \dontrun{
options(habistats.iucn_source = "path/to/your/data")
habistats::iucn_source_path()
} # }
old = options(habistats.cache_dir = "~/.cache/habistats-example")

sn = c("Strato arctica", "Diva satanica")
habistats::evaluate_iucn_range(sn)
#> # A tibble: 2 × 8
#>   sci_name       perimeter       area abundance richness idx_shannon idx_simpson
#>   <chr>              <dbl>      <dbl>     <int>    <int>       <dbl>       <dbl>
#> 1 Strato arctica  3201738.    3.39e11      9182        4       0.503       0.700
#> 2 Diva satanica   6784114.    8.04e11     18060        6       0.940       0.441
#> # ℹ 1 more variable: idx_brillouin <dbl>

options(old) # reset for this example, not needed in real use