Apply a function to each subset of a list.

lmap*() are similar to lapply(), but they map over subset .x[i] instead of element .x[[i]]. It allows .f to access the attributes of the encapsulating list. This is useful for operation on the sfc column in an sf data.frame. purrr::lmap() does similar things, but it forces the output to be a list.

mclmap*() are parallelized versions of lmap*().

Usage

lmap_lst(.x, .f, mode = c("list", "integer", "double", "character", "logical"))

lmap_lgl(.x, .f)

lmap_int(.x, .f)

lmap_dbl(.x, .f)

lmap_chr(.x, .f)

lmap_vec(.x, .f)

mclmap_lst(.x, .f, ...)

mclmap_lgl(.x, .f, ...)

mclmap_int(.x, .f, ...)

mclmap_dbl(.x, .f, ...)

mclmap_chr(.x, .f, ...)

mclmap_vec(.x, .f, ...)

Arguments

.x

A list.

.f

A function to apply to each subset of .x.

mode

Output type.

...

Additional arguments passed to parallel::mclapply().

See also