Rowwise operations

rowwise_*() functions apply a function to each row of a data.frame. The input to the function .f is a single-row subset with the same class. This is useful for handling subclasses of data.frame, such as sf. base::apply(.x, 1L, .f) cannot do this because it converts the input to a matrix.

rowwise_mcmap*() are parallelized versions of rowwise_map*().

Usage

rowwise_map(
  .x,
  .f,
  mode = c("list", "integer", "double", "character", "logical")
)

rowwise_map_lgl(.x, .f)

rowwise_map_int(.x, .f)

rowwise_map_dbl(.x, .f)

rowwise_map_chr(.x, .f)

rowwise_map_vec(.x, .f)

rowwise_mcmap(.x, .f, ...)

rowwise_mcmap_lgl(.x, .f, ...)

rowwise_mcmap_int(.x, .f, ...)

rowwise_mcmap_dbl(.x, .f, ...)

rowwise_mcmap_chr(.x, .f, ...)

rowwise_mcmap_vec(.x, .f, ...)

Arguments

.x

A data frame-like object.

.f

A function to apply to each row.

mode

Output type.

...

Additional arguments passed to parallel::mclapply().

See also